feat: move error handler folder
